Moved Navbar into a folder and LoginButton to its own file
