+++ autoscheduler/frontend/src/components/Navbar/NavBar.tsx
@@ -1,0 +1,57 @@
+import * as React from 'react';
+import {
+  AppBar, Toolbar, Typography, Button, makeStyles,
+} from '@material-ui/core';
+import { navigate } from '@reach/router';
+import LoginButton from './LoginButton';
+import appTheme from '../../theme';
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    flexGrow: 0,
+  },
+  menuButton: {
+    marginRight: theme.spacing(2),
+  },
+  title: {
+    flexGrow: 1,
+  },
+  buttonText: {
+    textTransform: 'capitalize',
+  },
+}));
+
+const NavBar: React.SFC = () => {
+  const classes = useStyles(appTheme);
+
+  return (
+    <div className={classes.root}>
+      <AppBar
+        position="static"
+        color="primary"
+      >
+        <Toolbar>
+          <div className={classes.title}>
+            <Button
+              color="inherit"
+              classes={{ label: classes.buttonText }}
+              onClick={(): void => {
+                // Navigate to the root when the title is clicked
+                navigate('/');
+              }}
+            >
+              <Typography variant="h6">
+                Rev Registration
+              </Typography>
+            </Button>
+          </div>
+          <Typography variant="subtitle1">
+            {LoginButton}
+          </Typography>
+        </Toolbar>
+      </AppBar>
+    </div>
+  );
+};
+
+export default NavBar;
